Conflicts:
	src/libgcm/materials/AnisotropicElasticMaterial.cpp
	src/libgcm/util/AnisotropicMatrix3DAnalytical.hpp
	src/tests/unit/anisotropic_matrixes.cpp

+++ src/libgcm/util/AnisotropicMatrix3DAnalytical.hpp
@@ -5,7 +5,10 @@
 #include "libgcm/util/MutableRheologyMatrix3D.hpp"
 #include "libgcm/util/matrixes.hpp"
 #include "libgcm/util/ThirdDegreePolynomial.hpp"
+#include "libgcm/util/MatrixInverter.hpp"
 #include "libgcm/Exception.hpp"
+#include <assert.h>
+
 #include <iomanip>
 
 using namespace gcm;
@@ -17,23 +20,23 @@ namespace gcm {
      * Creates corresponding rheology matrices for case of anisotropic material.
      * Implements 'semi-analytical' solution.
      */
-    class AnisotropicMatrix3DAnalytical : public MutableRheologyMatrix3D
-    {
-    public:
-        void createAx(const ICalcNode& node) override;
-        void createAy(const ICalcNode& node) override;
-        void createAz(const ICalcNode& node) override;
-    private:
-
-        void clear();
-        void fixValuesOrder();
-        void findNonZeroSolution(double **M, float *x);
-        void findNonZeroSolution(double **M, float *x, float *y);
-        void findEigenVec(float *eigenVec,
-                    double l, float rho, const IAnisotropicElasticMaterial::RheologyParameters &C, int stage);
-        void findEigenVec(float *eigenVec1, float *eigenVec2,
-                    double l, float rho, const IAnisotropicElasticMaterial::RheologyParameters &C, int stage);
-    };
+	class AnisotropicMatrix3DAnalytical : public MutableRheologyMatrix3D
+	{
+	public:
+		void createAx(const ICalcNode& node) override;
+		void createAy(const ICalcNode& node) override;
+		void createAz(const ICalcNode& node) override;
+	private:
+		
+		void clear();
+		void fixValuesOrder();
+		void findNonZeroSolution(double **M, double *x);
+		void findNonZeroSolution(double **M, double *x, double *y);
+		void findEigenVec(double *eigenVec,
+					double l, float rho, const IAnisotropicElasticMaterial::RheologyParameters &C, int stage);
+		void findEigenVec(double *eigenVec1, double *eigenVec2,
+					double l, float rho, const IAnisotropicElasticMaterial::RheologyParameters &C, int stage);
+	};
 }
 
 #endif
